+++ intelmq/tests/bots/parsers/autoshun/test_parser.py
@@ -30,7 +30,7 @@ EXAMPLE_EVENT0 = {
 EXAMPLE_EVENT1 = {
     "__type": "Event",
     "feed.name": "Autoshun",
-    "classification.type": "unknown",
+    "classification.type": "undetermined",
     "feed.url": "https://www.autoshun.org/files/shunlist.html",
     "raw": "PHRyPjx0ZD4xOTguNTEuMTAwLjg2PC90ZD48dGQ+MjAxNi0wNS0xOCAxNDowMTowNjwvdGQ+PHRkPk1TIFRlcm1pbmFsIFNlcnZlciBTaW5nbGUgQ2hhcmFjdGVyIExvZ2luLCBwb3NzaWJsZSBNb3J0byBpbmJvdW5kPC90ZD48L3RyPg==",
     "event_description.text": "MS Terminal Server Single Character Login, possible Morto inbound",